Add favicon.
"Jay, we don't need to focus on this right now, what are you doing?" Look, the favicon missing has been bothering me while I do work on this project. We're going to get this done and this was necessary for my sanity. Spare me the lectures while I do my work.